HSEARCH-792 - Consider renaming LockFactoryFactory to LockFactoryProvider
