+++ hibernate-search/src/main/java/org/hibernate/search/store/LockFactoryProvider.java
@@ -21,23 +21,32 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.hibernate.search.test.directoryProvider;
+package org.hibernate.search.store;
 
 import java.io.File;
 import java.util.Properties;
 
 import org.apache.lucene.store.LockFactory;
-import org.apache.lucene.store.SingleInstanceLockFactory;
-import org.hibernate.search.store.LockFactoryFactory;
 
-public class CustomLockFactoryFactory implements LockFactoryFactory {
-
-	// A real implementation would probably not use a static field; useful to keep the test simple.
-	static String optionValue;
-
-	public LockFactory createLockFactory(File indexDir, Properties dirConfiguration) {
-		optionValue = dirConfiguration.getProperty( "locking_option" );
-		return new SingleInstanceLockFactory();
-	}
+/**
+ * To use a custom implementation of org.apache.lucene.store.LockFactory
+ * you need to implement this interface and define the fully qualified
+ * classname of the factory implementation as a DirectoryProvider parameter
+ * for the locking_strategy key.
+ * The implementation must have a no-arg constructor.
+ *
+ * @author Sanne Grinovero
+ */
+public interface LockFactoryProvider {
+	
+	/**
+	 * Creates a LockFactory implementation.
+	 * A different LockFactory is created for each DirectoryProvider.
+	 * @param indexDir path to the indexBase setting, or null for
+	 * DirectoryProviders which don't rely on filesystem
+	 * @param dirConfiguration the properties set on the current DirectoryProvider
+	 * @return the created LockFactory
+	 */
+	LockFactory createLockFactory(File indexDir, Properties dirConfiguration);
 
 }
